+++ internal/dred/header.go
@@ -11,20 +11,20 @@ var errInvalidHeader = errors.New("dred: invalid experimental header")
 // Header mirrors the low-cost metadata libopus decodes before the model-heavy
 // DRED processing stage.
 type Header struct {
-	Q0              int
-	DQ              int
-	QMax            int
-	ExtraOffset     int
-	DredOffset      int
-	DredFrameOffset int
+	Q0              int32
+	DQ              int32
+	QMax            int32
+	ExtraOffset     int32
+	DredOffset      int32
+	DredFrameOffset int32
 }
 
-var dqTable = [...]int{0, 2, 3, 4, 6, 8, 12, 16}
+var dqTable = [...]int32{0, 2, 3, 4, 6, 8, 12, 16}
 
 // OffsetSamples converts the parsed DRED offset from 2.5 ms units to samples
 // at the caller's sample rate.
 func (h Header) OffsetSamples(sampleRate int) int {
-	return h.DredOffset * sampleRate / 400
+	return int(h.DredOffset) * sampleRate / 400
 }
 
 // EndSamples mirrors opus_dred_parse()'s dred_end output: the number of
@@ -48,7 +48,7 @@ func ParseHeader(payload []byte, dredFrameOffset int) (Header, error) {
 
 // QuantizerLevel mirrors libopus compute_quantizer() for the parsed DRED
 // quantizer metadata.
-func (h Header) QuantizerLevel(i int) int {
+func (h Header) QuantizerLevel(i int) int32 {
 	if i < 0 {
 		i = 0
 	}
@@ -56,10 +56,10 @@ func (h Header) QuantizerLevel(i int) int {
 	if dq < 0 {
 		dq = 0
 	}
-	if dq >= len(dqTable) {
-		dq = len(dqTable) - 1
+	if dq >= int32(len(dqTable)) {
+		dq = int32(len(dqTable) - 1)
 	}
-	quant := h.Q0 + (dqTable[dq]*i+8)/16
+	quant := h.Q0 + (dqTable[dq]*int32(i)+8)/16
 	if quant > h.QMax {
 		return h.QMax
 	}
